+++ packages/resolve/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @abux/resolve
 
+## 1.1.0
+
+### Minor Changes
+
+- Add supports for `pnpm`
+- Add yarn resolving fallback to pnpm or node_modules resolving (see yarn config `nodeLinker`)
+- Add debug mode
+
 ## 1.0.6
 
 ### Patch Changes